requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
python-multipart>=0.0.9
